requirements.txt
asgiref==3.6.0
attrs==20.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
Django==4.1.7
djangorestframework==3.14.0
idna==3.4
jellyfish==0.9.0
lib50==3.0.5
markdown2==2.4.8
pexpect==4.8.0
ptyprocess==0.7.0
pycparser==2.21
pytz==2023.3
PyYAML==5.4.1
requests==2.28.2
sqlparse==0.4.3
submit50==3.1.1
termcolor==1.1.0
urllib3==1.26.15